Unauthorize checking

+++ src/services/setup/CourseServices.js
@@ -1,5 +1,6 @@
 import axios from "axios";
 import config from "../../configuration/config";
+import { utils } from "../../helpers";
 
 let headers = {
     "Content-Type": "application/json",
@@ -8,6 +9,8 @@ let headers = {
 export const CourseServices = {
     fetchCategory: async () => {
         try {
+            const hasPermission = utils.hasPermission("course","read");
+            if(!hasPermission) throw new Error("You do not have permission.");
             const result = await axios.get(config.courseCategoryUrl, { headers: headers });
             console.log('Categories From Service : ', result);
             return result.data;
@@ -24,6 +27,8 @@ export const CourseServices = {
     createCategory: async (payload) => {
         // headers.Authorization = localStorage.getItem("_token");
         try {
+            const hasPermission = utils.hasPermission("course","create");
+            if(!hasPermission) throw new Error("You do not have permission.");
             const result = await axios.post(config.otherCourseCategoryUrl, payload, { headers: headers });
             if(result.status === 200) {
                 return result.data;
@@ -43,6 +48,8 @@ export const CourseServices = {
     editCategory: async (payload) => {
         // headers.Authorization = localStorage.getItem("_token");
         try {
+            const hasPermission = utils.hasPermission("course","update");
+            if(!hasPermission) throw new Error("You do not have permission.");
             const url = config.otherCourseCategoryUrl + "/" + payload.id;
             const result = await axios.put(url, payload, { headers: headers });
             if(result.status === 200) {
@@ -63,6 +70,8 @@ export const CourseServices = {
     deleteCategory: async (catID) => {
         // headers.Authorization = localStorage.getItem("_token");
         try {
+            const hasPermission = utils.hasPermission("course","delete");
+            if(!hasPermission) throw new Error("You do not have permission.");
             const url = config.otherCourseCategoryUrl + "/" + catID;
             const result = await axios.delete(url, { headers: headers });
             if(result.status === 200) {
@@ -82,6 +91,8 @@ export const CourseServices = {
 
     fetchCourses: async (page=null, queryParams=null) => {
         try {
+            const hasPermission = utils.hasPermission("course","read");
+            if(!hasPermission) throw new Error("You do not have permission.");
             const result = await axios.get(config.fetchCourseUrl +
             (page ? "/" + page : '') +
             (queryParams ? "?" + decodeURI(queryParams) : ''),
@@ -101,6 +112,8 @@ export const CourseServices = {
 
     fetchCourseList: async (catID) => {
         try {
+            const hasPermission = utils.hasPermission("course","read");
+            if(!hasPermission) throw new Error("You do not have permission.");
             const url = config.fetchCourseUrl + "/1?catID=" + catID;
             const result = await axios.get(url, 
             { headers: headers });
@@ -147,6 +160,8 @@ export const CourseServices = {
     createCourse: async (payload) => {
         // headers.Authorization = localStorage.getItem("_token");
         try {
+            const hasPermission = utils.hasPermission("course","create");
+            if(!hasPermission) throw new Error("You do not have permission.");
             const result = await axios.post(config.createCourseUrl, payload, { headers: headers });
             if(result.status === 200) {
                 return result.data;
@@ -166,6 +181,8 @@ export const CourseServices = {
     editCourse: async (payload) => {
         // headers.Authorization = localStorage.getItem("_token");
         try {
+            const hasPermission = utils.hasPermission("course","update");
+            if(!hasPermission) throw new Error("You do not have permission.");
             const url = config.editCourseUrl + "/" + payload.id;
             const result = await axios.put(url, payload, { headers: headers });
             if(result.status === 200) {
@@ -186,6 +203,8 @@ export const CourseServices = {
     deleteCourse: async (productID) => {
         // headers.Authorization = localStorage.getItem("_token");
         try {
+            const hasPermission = utils.hasPermission("course","delete");
+            if(!hasPermission) throw new Error("You do not have permission.");
             const url = config.deleteCourseUrl + "/" + productID;
             const result = await axios.delete(url, { headers: headers });
             if(result.status === 200) {
